refactor(app): outline `Dns` metrics into submodule
